LayoutTests/fast/workers/worker-script-error.html

<body>
<p>Test Worker script error handling functionality. Should print a series of PASS messages, followed with DONE.</p>
<div id=result></div>
<script>
function log(message)
{
    document.getElementById("result").innerHTML += message + "<br>";
}

var testCases = [
    "testInvalidScriptSyntax",
    "testScriptErrorUnhandled",
    "testScriptErrorAddEventListener",
    "testErrorInHandlingScriptError",
    "testScriptErrorBubbledAndHandledInWorker",
    "testScriptErrorBubbledAndReportedToUser",
    "testScriptErrorHandled",
];
var testIndex = 0;

function runNextTest()
{
    if (testIndex < testCases.length) {
        testIndex++;
        window[testCases[testIndex - 1]]();
    } else {
        log("DONE");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }
}

function testInvalidScriptSyntax()
{
    try {
        var worker = new Worker("resources/worker-invalid-syntax.js");
        worker.onerror = function() {
            log("PASS: onerror invoked for a script that has invalid syntax.");
            runNextTest();
            return false;
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testScriptErrorUnhandled()
{
    try {
        var worker = new Worker("resources/worker-script-error-unhandled.js");
        worker.onerror = function(evt) {
            log("PASS: onerror invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
            runNextTest();
            return false;
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testScriptErrorAddEventListener()
{
    try {
        var worker = new Worker("resources/worker-script-error-unhandled.js");
        worker.addEventListener("error",
            function(evt) {
                log("PASS: event listener invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
                runNextTest();
                evt.preventDefault();
            }, false);
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testErrorInHandlingScriptError()
{
    try {
        var worker = new Worker("resources/worker-error-in-handling-script-error.js");
        worker.onerror = function(evt) {
            log("PASS: onerror invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
            runNextTest();
            return false;
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testScriptErrorBubbledAndHandledInWorker()
{
    try {
        var worker = new Worker("resources/worker-script-error-bubbled.js");
        worker.onerror = function(evt) {
            log("PASS: onerror invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
            runNextTest();
            return false;
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testScriptErrorBubbledAndReportedToUser()
{
    try {
        var worker = new Worker("resources/worker-script-error-bubbled.js");
        worker.onerror = function(evt) {
            log("PASS: onerror invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
            runNextTest();
            return true;
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

function testScriptErrorHandled()
{
    try {
        var worker = new Worker("resources/worker-script-error-handled.js");
        worker.onerror = function(evt) {
            log("FAIL: onerror invoked for a script that has script error '" + evt.message + "' at line " + evt.lineno + ".");
            runNextTest();
            return false;
        }
        worker.onmessage = function(evt) {
            log("PASS: message received from WorkerGlobalScope.onerror: " + evt.data + ".");
            runNextTest();
        }
    } catch (ex) {
        log("FAIL: unexpected exception " + ex);
        runNextTest();
    }
}

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

runNextTest();

</script>
</body>
