LayoutTests/fast/workers/worker-script-error-expected.txt

CONSOLE MESSAGE: line 7: ReferenceError: Can't find variable: foo
Test Worker script error handling functionality. Should print a series of PASS messages, followed with DONE.

PASS: onerror invoked for a script that has invalid syntax.
PASS: onerror invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 1.
PASS: event listener invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 1.
PASS: onerror invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 7.
PASS: onerror invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 7.
PASS: onerror invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 7.
PASS: message received from WorkerGlobalScope.onerror: onerror invoked for a script that has script error 'ReferenceError: Can't find variable: foo' at line 7.
DONE
